requirements.txt
streamlit
pandas
numpy
py3Dmol
stmol
streamlit-aggrid
Pillow